In class TestComplexFigure: For the purpose of testing, we will create single object of Circle and use it as the argument of constructor for instance of ComplexFigure-class.      For creating objects and catching related exceptions we will use factory method in special class FigureFactory.      Next, we will create a pair of GeomFigure-child and use it in call of method addFigure(). So, we have an instance of ComplexFigure-class containing triangle, circle and parallelogram objects.      Now, lets call methods for calculating summary of perimeters of it all, and methods for calculating summary of areas of this objects.      Make factory for creating objects (reference to dependency injection ;) )

+++ src/org/dng/Parallelogram.java
@@ -11,7 +11,7 @@ public class Parallelogram extends GeomFigure {
     private double b;
     private int angle;
 
-    public Parallelogram(int a, int b, int angle) throws Exception {
+    public Parallelogram(double a, double b, int angle) throws Exception {
         if ((a <= 0) || (b <= 0) || (angle <= 0) )
             throw new Exception("Parallelogram with such parameters cant exist!");
